requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
zipp==3.17.0

pydantic~=2.5.3
SQLAlchemy~=2.0.23
alembic~=1.13.1
python-dotenv~=1.0.0
PyJWT~=2.8.0